Rename folder and imports for SnapPermissionCell
