+++ ui/components/app/snaps/snap-permission-cell/index.js
