+++ ui/components/app/snaps/snap-permission-cell/snap-permission-cell.js
