+++ ui/components/app/snaps/snap-permissions-list/snap-permissions-list.js
@@ -9,7 +9,7 @@ import {
   getSnapsMetadata,
 } from '../../../../selectors';
 import { getSnapName } from '../../../../helpers/utils/util';
-import SnapPermissionCell from '../snap-permission-cell-display';
+import SnapPermissionCell from '../snap-permission-cell';
 
 export default function SnapPermissionsList({
   snapId,
